Make build work on Linux with gcc
timer_t is already defined in time.h on Linux and other unices that use a later POSIX version than MacOS, so I've renamed the local type.

gcc doesn't like to cast a const char* to a char* and that was never a guarantee anyway so I changed the type of the local variable.